+++ changelog.md
@@ -2,6 +2,7 @@
 - Fixed multiblock rendering in the manual with Optifine and texture packs (Malte)
 - Fixed pre-IE-79 multiblocks not working on IE 79 (Malte)
 - Fixed FluidContainer slots not accepting empty containers (BluSunrize)
+- Fixed Villager trades for Shader bags being inverted (BluSunrize)
 
 #####Version 0.12-79 - BUILT
 - Added the ability for Wirecutters to cut wires from anywhere (Malte)